+++ pyreference/settings.py
@@ -3,14 +3,9 @@
 PYREFERENCE_JSON_VERSION_KEY = "pyreference_json_version"
 
 # Keys used in dictionary (serialized to JSON)
-CHROM = "chr"
+CONTIG = "contig"
 START = "start"
 END = "stop"
 STRAND = "strand"
 
-# Other 
-IS_CODING = "is_coding"
-
-
 BEST_REGION_TYPE_ORDER = ["coding", "5PUTR", "3PUTR", "non coding", "intron"]
-